WebNN: a single big CL to prototype GPU backend
Implement WebNN GPU backend in the GPU process with a singleton service which
communicates with render process with mojo infrastructure. There are a WireServer
in GPU and WireClient in render process for every navigator.ml. The Object ID of
each IDL interface is used to refer to this object when talking with the server side.

This prototype supports 8 operations required by MobileNet V2 [1] for performance
evaluation purpose. The input and output data use shared buffer across process.
WebNN DirectML is about 260% faster than WebGL on Windows platform.

The CL is based on patch 18 of XNNPACK's CL.

[1]: https://arxiv.org/abs/1801.04381

Bug: 1273291