WebNN: prototype WebNN implementation on ChromeOS platform
This prototype demonstrates a design proposal that runs WebNN computational
graphs in MLService to access hardware acceleration by using ModelLoader mojom
interface. In this approach, the WebNN graph (MLOperators and MLOperands) would
be converted into a TF-Lite model in FlatBuffers with TF-Lite schema, in
renderer side. The MLGraphBuilder::build() would be implemented by feeding the
converted TFLite FlattBuffers to the load() method [1] of ModelLoader mojo
interface, The compute() will be implemented by invoking the compute() method
[2] of Model mojo interface.

[1] https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform2/ml/mojom/web_platform_model.mojom;l=140
[2] https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform2/ml/mojom/web_platform_model.mojom;l=152

Bug: 1273291