WebNN: Define average and max pool2d operator in mojo.
The pool2d is to compute a mean or max reduction operation across all
the elements within the moving window over the input tensor.

Validate the pool2d operators with converting the attributes from mojo
to component library type and add unit tests to verify to the graph
from renderer process.

Bug: 1273291