WebNN: Define async computation in WebNNGraph mojo interface.
The `Compute` method is called by the renderer process to carry out the
computational workload of the compiled graph. All array buffers of
inputs/outputs for computing graph are shared in one shared memory
between renderer process and GPU process, the `ArrayBufferInfo` mojo
struct is used to describe the information of each chunk in the memory.

Bug: 1273291