[WebNN][CrOS]: Support Conv2d and Pool2d operators
Conv2d is a 2-D convolution given the input and filter tensors, and
only nhwc input layout and partial activation are supported at current
stage, serialize a empty buffer with the size of output channel if
there is no bias operand.

Pool2d is an average or max pooling operation across all the elements,
and only nhwc layout is supported.

Bug: 1273291